pyproject.toml
[project]
name = "flask-sample-app"
version = "1.0.0"
dependencies = [
    "Flask",
    "dotenv-cli",
    "Flask-SQLAlchemy",
    "Flask-Migrate",
    "argon2-cffi",
    "Flask-WTF",
    "invoke",
    "gunicorn",
    "uvicorn",
    "ruff",
    "black",
]

[tool.black]
line-length = 80